healthcheck.sh add --no-connect option
There is an implicit --connect option in healthcheck
so that a normal use can be sure MariaDB is running
on a tcp socket.

There is a case in /docker-entrypoint-initdb.d where
its desirable to perform healthchecks for components
without a --connect. In this case, use --no-connect
in the healthcheck to avoid an implicit --connect test.